Make sure to load dependencies first
Dir.glob does not guarantee a particular order in which the entries are
returned. Depending on the machine the "shellout_type_assertions.rb"
entry might be returned after files that depend on it which leads to an
error.

This is now fixed by loading it explicitly before loading all the other
files in "support/".