Relax constraint on hostname
With xenial on GCP we observe that the hostname gets resolved to the
internal vm id instead of the BOSH agent id. Since `localhost` maps to
the BOSH agent id in `/etc/hosts`, the syslog format does not resolve
`hostname` to localhost. Thus we needed to relax the tests assertion.

[#157863470](https://www.pivotaltracker.com/story/show/157863470)

Co-authored-by: James Myers <jmyers@pivotal.io>