Remove BOSH libyaml stage
This change removes the bosh_libyaml stage, which installs a locked
version of libyaml on the stemcell. We think that this was added to
support CAPI who at one point required libyaml to be present to work.
According to https://www.pivotaltracker.com/story/show/128550439, we
believe that this is no longer necessary.

This installation process was also conflicting with the `libyaml-devel`
installation on Centos 7.5, required by openSSL.

[#158311224](https://www.pivotaltracker.com/story/show/158311224)

Signed-off-by: Florian Nachtigall <florian.nachtigall@sap.com>