Bump deboostrap to 1.0.102
It seems we've bumped this past 1.0.49 because of reasons(*). But then
kept bumping forward when other versions are deleted from the
repository. debootstrap_1.0.78+nmu1ubuntu1.5_all.deb is now gone so we
picked the absolute latest available package (1.0.102).

(*): We didn't really go that deep into the reasons there, but they can
be found in a comment on apply.sh here and on the commits in the bosh
repo that added that change.

[#158377910](https://www.pivotaltracker.com/story/show/158377910)

Co-authored-by: Luan Santos <lsantos@pivotal.io>