Make bosh-agent CPU priority higher than BOSH jobs
We want the bosh-agent to continue communicating with the BOSH Director,
even then the instance is genuinely CPU-contended.

If the bosh-agent has the same CPU priority as all BOSH jobs,
when the instance is under extended CPU pressure,
it has been observed for the HM to re-create the host due to the
bosh-agent stopping communication with the Director.
This is bad, because the CPU is genuinely contended, and
re-creating the instance makes matters worse,
especially for distributed stateful systems such as RabbitMQ.

This was first reported as
https://github.com/cloudfoundry/bosh/issues/1949
and acknowledged by @cppforlife.