Autobump bosh-agent in xenial master
We are using meta4 to automatically bump the bosh-agent.  The
bosh-agent-index git repository is pushed to whenever the bosh-agent
pipeline is complete and the new file that is added is a new bosh-agent
version.  We use the name of the metalink file to discover the
bosh-agent version number and we use the meta4 CLI to download that
agent using the meta4 metalink file format that was committed in the
latest bosh-agent git repository commit.

[#156236052](https://www.pivotaltracker.com/story/show/156236052)

Co-authored-by: Joshua Aresty <joshua.aresty@emc.com>