Correctly remove google_set_hostname script
Previously, on trusty, the hostname was configured in such a way to be
the agent id. When we switched over to xenial, we intended to keep the
same precedent. However, upstream, Google decided to rename one of their
underlying vm bootstrapping scripts, which conveniently caused racy
conditions that would occasionally change the hostname back to the vm
cid. This commit ensures that the renamed script will not run by
removing it.

[#162225262](https://www.pivotaltracker.com/story/show/162225262)

Co-authored-by: Belinda Liu <bliu@pivotal.io>