Run builder with host UID and GID
The stemcell builders now require their Docker images to be passed the
host's UID and GID. The reason for this is because the container has a
bind mounted directory shared with the host.

This commit ensures that the container is built with a user that has the
same UID and GID as the host, satisfying the file permissions of the
bind mounted directory.