Fix locale generation again
This commit defines a locale file using the `localedef` command because
the error was complaining about en_US.UTF-8 not existing. It also
ensures that the `dpkg-reconfigure` runs in noninteractive mode.
Previously this command was running with the default input method
causing it to pick a default locale (not what we want -- en_US.UTF-8).