WIP this work was part of an exploration into using the linux-aws ubuntu package as a potential fix for the flappy netwwork issue #99
With the changes in this commit a ubuntu xenial stemcell can be build
with the linux-aws kernel. However when bosh create a vm with this
stemcell it fails to boot (gets stuck in the bootloader).

Since this spike was time boxed we unfortunately no time left to further
explore this GRUB issue.

https://www.pivotaltracker.com/story/show/174635234