Force rsyslog to wait for /var/log to be mounted
As part of the bosh-agent start up process, it mounts /var/vcap to the
persistent disk attached to the VM. It is intended for rsyslog to start
after this happens. After making some changes to the packages installed
on the GCP stemcell, rsyslog was starting before /var/vcap was properly
mounted.

This change adds a pre-start script to the rsyslog service that forces
it to wait until /var/vcap is properly mounted.

[#177163652] Investigate cf ssh related failures

Signed-off-by: Kenneth Lakin <klakin@vmware.com>