Revert "Revert "Flip test for stemcell image having 2 partitions""
Our understanding the logic was flawed, and the initial logic was
correct. We are reverting.

The confusion stems from the rspec filter, `:exclude_on_softlayer`,
which means one thing in testing, another in building.

In building, which is what this change affects, `:exclude_on_softlayer`
set to `true` means that we're building the SoftLayer stemcell. In
rspec, it means we're _not_ testing the Softlayer stemcell.

This reverts commit 0a0258b897c149fa193dfdbabf3ec1b83de920be.

fixes <https://main.bosh-ci.cf-app.com/teams/main/pipelines/bosh:stemcells:ubuntu-bionic/jobs/build-google-kvm-master/builds/47>:
```
  2) Ubuntu 18.04 OS image installed by system_grub /boot/grub/unicode.pf2 should be file
     Failure/Error: it { should be_file }
       expected `#<ShelloutTypes::File:0x000055e9551731a0 @path="/boot/grub/unicode.pf2", @chroot=#<ShelloutTypes::Chroot:0x000055e9551731c8>>.file?` to return true, got false
     # ./spec/os_image/ubuntu_bionic_spec.rb:118:in `block (5 levels) in <top (required)>'

```

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>

[#149]
[#177200369] cloudfoundry/bosh-linux-stemcell-builder #149: Please revert commit "Flip test for stemcell image having 2 partitions"