Machine-id is unique for each created VM
Per the [GitHub issue](https://github.com/cloudfoundry/bosh-linux-stemcell-builder/issues/122):

> The duplication of the machine-id causes instances to use the same
dhcp client identifier when running a dhcp client (e.g.
systemd-networkd). This in turn causes dhcp-servers get dhcp-discovers
from multiple different VMs with the same client-identifer. Because of
that the system can shortly loose it's ip-address which can cause
various issues (similar to this).

The initial work (for Ubuntu Bionic) was done by Stefan Wutz, and we've
extended this to Ubuntu Xenial.

[#177701034] cloudfoundry/bosh-linux-stemcell-builder #155: Machine-id should be unique for each created vm

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>