Add rsyslog-mmnormalize pkg, allow for better filtering
We'd like to add rsyslog-mmnormalize package to allow for better filtering. As syslog-release of cf uses the OS rsyslogd the current way to go is to add it here.
This module is needed to properly filter the rather unconventional cloud_controller_ng syslogs format.
Usecase: We have is to just relay syslog messages for audit log purposes that contain altering http operations (PUT/DELETE etc.) and do not forward normal GET requests.