Add stage to build a Ubuntu Bionic FIPS stemcell
There are usecases where FIPS compliant stemcells are needed. So
adding an option to be able to build FIPS compliant stemcells here.

Note: The implementation depends on the UBUNTU_ADVANTAGE_TOKEN
environment variable. If that variable is set (it needs to be a valid
token because it's later used by the ubuntu-advantage tool to install
the relevant FIPS packages), the build will produce a FIPS stemcell.