Build a Ubuntu Bionic Pro FIPS stemcell variant
When building an image with something like:

$ bundle exec rake --trace stemcell:build_os_image[ubuntu,bionic-fips,\
  $PWD/tmp/ubuntu-bionic-fips.tgz]

the operating system variant will be 'fips' (see commit 70c4e0d6dab6).
Based on that variant, include the newly added "base_apt_fips" stage
to make the image FIPS compliant.

Note: a UBUNTU_ADVANTAGE_TOKEN needs to be provided as environment
variable to make the build working.

Note: when using the built FIPS stemcell, a Ubuntu Advantage[0]
subscription is needed for every machine the stemcell is running on.

[0] https://ubuntu.com/advantage