Rakefile: Fix ArgumentError errors from FileUtils method calls.
NOTE: All of the info below also seems to be related to differences in behavior between Ruby 2.x vs 3.x. Not sure exactly what the root cause of the difference or incompatibility is, though.

NOTE: The previous code seems like it _should_ have been fine, but the unprefixed `mkdir_p` calls were raising "ArgumentError: wrong number of arguments (given 2, expected 1)" errors.

According to the following links, it seems like the unprefixed `mkdir_p` calls _should_ have worked fine:
- https://github.com/ruby/rake/blob/master/lib/rake/dsl_definition.rb#L16-L24
- https://github.com/ruby/rake/blob/master/lib/rake/file_utils_ext.rb#L10-L11

However, changing the failing method calls to `FileUtils.mkdir_p` resolves the errors. It seems possible that the errors are related to the following line and commit, which changed the mkdir_p method's signature.
- https://github.com/ruby/fileutils/blame/master/lib/fileutils.rb#L206
- https://github.com/ruby/fileutils/commit/482de6d397742526d1111576e2791f9b7051e3c0

So it would seem like something about the way that the Rake dsl_definition.rb is including the FileUtils methods in the Rake DSL's namespace is resulting in the unprefixed method invocations behaving differently than the invocations which explicitly call the FileUtils module's methods directly. Not sure why. But it seems like these changes are unlikely to have any unintended side effects, and all of the other calls to `mkdir_p` within the project already seem to use the explicit prefix.