🐞 `rake stemcell:build_os_image` specs succeed
The tests requiring the allocation of a pty (pseudo-tty) would fail when
the `/dev/pts` directory (necessary for ptys) was obscured by an
ill-thought bind mount of `/dev` to itself.

We no longer bind mount `/dev` to itself, but we do allow it
to be bind mounted to another directory (specified by `@os_image_dir`).

We are unsure why this problem has suddenly manifested, for this section
of code hadn't changed in a long time. Perhaps a change in the version
of `sudo`, of `rspec`, or test setup?

Fixes, during `bundle exec rake stemcell:build_os_image[ubuntu,jammy,$PWD/tmp/ubuntu_base_image.tgz]`:
```
sudo: unable to allocate pty
```