rhel: stages: Apply tty_config stage to all RHEL stemcells.
NOTE: Based on the fact that spec examples which depend on the actions of the tty_config stage already existed in os_image_shared_examples.rb prior to the original removal of Centos/RHEL support (by 6acd0c9c), it would be assumed that all Centos and RHEL stemcells would need to run that stage. However, the stage was previously included only for Centos 7 and Ubuntu. Now, it is also included as part of the RHEL 7 & 8 stages.

See (c54d5a3c is the parent commit of 6acd0c9c):
- https://github.com/cloudfoundry/bosh-linux-stemcell-builder/blob/c54d5a3c6ff8856c85567b03a669ef8ff387bd80/bosh-stemcell/spec/support/os_image_shared_examples.rb#L84-L96
- https://github.com/cloudfoundry/bosh-linux-stemcell-builder/blob/c54d5a3c6ff8856c85567b03a669ef8ff387bd80/bosh-stemcell/lib/bosh/stemcell/stage_collection.rb#L256

Fixes/resolves the following spec failures:

> Failed examples:
> ...
> rspec './spec/os_image/rhel_8_spec.rb[1:1:7:2:1:1:1]' # RHEL 8 OS image behaves like every OS image /etc/securetty restricts root login to system console (CIS-9.4) awk '$1 !~ /^(console|#.*| *)$/ { print; f=1 } END { if (!f) print "none" }' /etc/securetty stdout is expected to eq "none\n"