rhel: stages: Apply escape_ctrl_alt_del stage to all RHEL stemcells.
NOTE: Based on the fact that spec examples which depend on the actions of the escape_ctrl_alt_del stage already existed in os_image/centos_7_spec.rb (but NOT in os_image/rhel_7_spec.rb) prior to the original removal of Centos/RHEL support (by 6acd0c9c), it would be assumed that either that stage AND the corresponding specs were either both intentionally omitted for RHEL, OR that they should both be applied to RHEL (for symmetry with Centos). This commit implements the latter (i.e. apply the stage and specs to RHEL).

See (c54d5a3c is the parent commit of 6acd0c9c):
- https://github.com/cloudfoundry/bosh-linux-stemcell-builder/blob/c54d5a3c6ff8856c85567b03a669ef8ff387bd80/bosh-stemcell/spec/os_image/centos_7_spec.rb#L171-L176
- https://github.com/cloudfoundry/bosh-linux-stemcell-builder/blob/c54d5a3c6ff8856c85567b03a669ef8ff387bd80/bosh-stemcell/lib/bosh/stemcell/stage_collection.rb#L261

Fixes/resolves the following spec failures:

> Failed examples:
> ...
> rspec ./spec/os_image/rhel_8_spec.rb:165 # RHEL 8 OS image overriding control alt delete (stig: V-38668) /etc/systemd/system/ctrl-alt-del.target remarks on the escaping