🐞 Don't try to purge `crda` package; it's not there
Similarly, don't try to purge the `crda-wireless` package; it's not
there either.

"crda is the Linux wireless central regulatory domain agent." For
example, it ensures that WiFi devices locate within the USA comply with
FCC restrictions.

Fixes, when creating os-image:
```
Package 'crda' is not installed, so not removed
E: Unable to locate package wireless-crda
```