Fix specs: Bosh::Stemcell::StageCollection#operating_system_stages when ... has the correct stages
NOTE: This fixes a spec failure introduced in 527fdd1c and related to a merge conflict between 527fdd1c and 2108b483 (see issue #122). The failure was undetected at the time as a result of issue #226.

NOTE: Based on the content of #122, it seems like the 'clean_machine_id' stage probably should remain an Ubuntu-specific stage. Therefore, it will not be applied to CentOS or RHEL stemcells at this time.

See:
- 527fdd1c: Revert "Wip: remove centos, rhel, suse, photonos, trusty"
- 2108b483: Machine-id should be unique for each created vm
- https://github.com/cloudfoundry/bosh-linux-stemcell-builder/issues/122
- https://github.com/cloudfoundry/bosh-linux-stemcell-builder/issues/226

Fixes spec failures:
- 0) Bosh::Stemcell::StageCollection#operating_system_stages when Ubuntu has the correct stages
- 0) Bosh::Stemcell::StageCollection#operating_system_stages when CentOS 7 has the correct stages