NATS access helper aids troubleshooting
Newest stemcells prevent processes from accessing NATS (port 4222) on
the BOSH Director because security; however, this can hamper
troubleshooting.

This commit adds a helper script to enable an operator to access the
NATS port on the director; it accomplishes this by adding the PID to the
cgroup which is allowed to communicate with the Director's NATS server.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>

[#182005633](https://www.pivotaltracker.com/story/show/182005633)