Accommodate `modprobe`'s new location in `/usr/bin`
The `syslog storer` tests assume that `modprobe` is located in `/sbin/`,
where it's been for at least a decade. Unfortunately that's no longer
true, and on Ubuntu Jammy Jellyfish it's located in `/usr/sbin/`. In
fact, on Jammy nothing is in `/sbin/` because it's linked to
`/usr/sbin/`. And, since we're on the topic of links, `modprobe` is
nothing but a link to `/bin/kmod`, which is the Grand Unified Kernel
Module Executable (aka "GUKME", a term which I've made up on the spot);
i.e. it's also where `lsmod`, `insmod`, etc. are linked to.

This commit fixes that faulty assumption by allowing both old & new
locations of `modprobe`.

Fixes <https://bosh.ci.cloudfoundry.org/teams/main/pipelines/stemcells-ubuntu-jammy/jobs/test-stemcells-master-ipv4/builds/87>:
```
[Fail] Syslogrelease when auditd is monitoring access to modprobe [It] gets forwarded to the syslog storer
/tmp/build/179b5d98/bosh-linux-stemcell-builder/src/github.com/cloudfoundry/stemcell-acceptance-tests/ipv4director/syslogrelease/smoke_test.go:21
```