Bump vSphere Stemcell virtualHW.version 9 → 13
- Updating compatibility adds new features (such as EVC-per-host) and
  mitigates certain security bugs, e.g. [Security Advisory
  VMSA-2018-004.3](https://www.vmware.com/security/advisories/VMSA-2018-0004.html)
  "Ensure that your VMs are using Hardware Version 9 or higher. For best
  performance, Hardware Version 11 or higher is recommended"
- We haven't updated the vSphere stemcell's `virtualHW.version` ("VM
  Compatibility") since 2018-01
- The previous compatibility was 9, "ESXi 5.1 and later"
- The new compatibility is 13, "ESXi 6.5 and later", which is the oldest
  supported version of ESXi:
    - ESXi 6.0's EOGS (end of general support) was March 12, 2020, and
      the EOTG (end of technical guidance) was March 12, 2022.
    - ESXi 6.5 (the next release, and the next compatibility) has an
      EOGS of October 15, 2022, and an EOTG of November 15, 2023

We also switched from the IDE drivers to the SATA drivers for the
CD-ROM. We haven't seen IDE cables in the wild in ten, maybe fifteen,
years. This is a purely cosmetic change in that it makes our VMs &
stemcells appear up-to-date but doesn't offer any new features.

Drive-by: we updated the README's Quick Start to not include a stemcell
version (letting it default to "0000"), which means we don't need to
keep bumping the stemcell version in the README as we cut new releases.