README: accommmodate M1/M2 macs
Stemcells can be built on an M1/M2 (ARM-based) Mac provided Rancher
Desktop is used and the appropriate flags are passed.

This commit updates the README with the appropriate flags based on the
underlying architecture (i.e. `uname -m` returns`arm64`, not the
Intel-based `x86_64`). Should developers decided to build stemcells on
AWS's ARM-based Graviton instances, these instructions should be updated
to account for an architecture of `aarch64`.

Drive-by: we ignore the `.vscode`, the directory created my Microsoft's
increasingly-popular editor, Visual Studio Code.