Revert "Configure chrony to use hardware clock on Azure"
The reverted commit added the functionality of syncing time against PHC
(Precision Hardware Clock) and prevented customers from adding network
rules to forward NTP traffic. However in ubuntu-xenial there are bugs in
the kernel that prevent PHC from working. It's unlikely to have a fix
for this on xenial. Because of that this revert brings back NTP to sync
the date time.

This reverts commit f320211548220ccb1efa63f22b59948fbed17b5f.