rsyslogd runs on Jammy
Prior to this commit, `rsyslogd`, the logging daemon, was not running on
Jammy-based stemcells even though it should be running.

We also had to make sure that `rsyslogd` was logging _all_ the logs, especially
the ones emitted by the systemd Journal (`journald` or `systemd-journald`).

A challenge we had to overcome was that `journald` stopped when the BOSH agent
mounted `/var/log` early in the boot process; our fix was to configure
`journald` to write to volatile storage (`/run/log/journal/`) instead of
persistent (`/var/log/journal`). This did not jeopardize the logs; `journald`
forwards its logs to `rsyslogd` via a socket. We also configure `rsyslogd` to
wait until `/var/log` has been mounted before starting.