Persistent disk resizing works after agent restart
Related issue: https://github.com/cloudfoundry/bosh/issues/2433

If the agent has been restarted, attempting to remount the persistent disk
as read-only result in the error:
  already mounted or mount point busy
And a kernel error of:
  Can't mount, would change RO state

We do not fully understand the root cause of the problem, but it the restarting
of the chrony service via systemd is the action triggering the problem.

Running chronyd's start script directly does not trigger the problem, nor
is the problem triggered by restarting other services via systemd.

Signed-off-by: Chris Selzo <cselzo@vmware.com>
Co-authored-by: Chris Selzo <cselzo@vmware.com>
Co-authored-by: Long Nguyen <nguyenlo@vmware.com>
Co-authored-by: Joseph Palermo <jpalermo@pivotal.io>