Add systemd config to prevent chronyd from locking bind mounts
Issue: https://github.com/cloudfoundry/bosh/issues/2433

Chrony's systemd config uses ReadWritePaths to limit the files it has access to.
When this is configured, systemd creates a file system namespace for the service.
This namespace ties up the existing bind mounts, even if the service isn't given
access to those particular paths. This can cause disk unmounting to fail if the
file namspace was created when that mounting existed.

This was most commonly seen when chrony had been restarted AFTER the
persistent disk had been mounted. Normally when doing a persistent disk
migration.

This changes adds back InaccessiblePaths to prevent the file namspace from
having access to /var/vcap/store.

This work was done previously in: https://github.com/cloudfoundry/bosh-linux-stemcell-builder/pull/152
but was lost as part of a later cleanup.