Changes logrotate.timer trigger to OnCalendar
This commit changes the logrotate Timer Unit's scheduling mechanism from
"OnUnitActiveSec" to "OnCalendar".

Why? A customer reported that on their Jammy-stemcell-based software,
logrotate was never being run. Support reproduced the problem in their
lab.

Investigation revealed that scheduling oneshot Timer Units with
"OnUnitActiveSec" is something that has been long-known to be
unsupported [1], and has had recent trouble reports from people running
software on Jammy [2].

As further evidence of this change's likely correctness, Canonical also
made a similar change for the same reason back in 2019. [3] (The
programmer making the fix tries a few different things, but eventually
discovers that "OnCalendar" works correctly. [4])

Unfortunately, this problem usually does not happen, and getting it TO
happen can be very difficult, so proving the correctness of this fix is
hard.

[1] <https://github.com/systemd/systemd/issues/6680#issuecomment-326228938>
[2] <https://github.com/systemd/systemd/issues/6680#issuecomment-1852339223>
[3] <https://bugs.launchpad.net/ubuntu/+source/base-files/+bug/1829968>
[4] <https://bugs.launchpad.net/ubuntu/+source/base-files/+bug/1829968/comments/5>