fix: don't change the systemd default from shared
If the default is private, then if namespaced services (like
auditd/chronyd) are restarted after the persistent disk is mounted, then
it is not possible to fully unmount that disk until those services are
stopped. The system will appear to let you do so, however the device
remains busy and subsequent attempts to remount (as read-only) will fail
as the device is busy.

Signed-off-by: Adam Eijdenberg <adam.eijdenberg@defence.gov.au>