Remove old metalink
This now seems to be stored in a path that includes a # for some reason.
We think the Concourse metalink resource was updated back in June of 2023
and that update broke the parsing in some way.

But we have been storing the metalink in the new path for a year now, so we are deleting
the old copy to reduce confusion

Signed-off-by: Joseph Palermo <joseph.palermo@broadcom.com>
Co-authored-by: Joseph Palermo <joseph.palermo@broadcom.com>