Lower ClientAliveInterval for ssh
The Azure Marketplace requires that images have set ClientAliveInterval
to between 30 and 235, with a recommended value of 180. It seems safe to
update the value for all IaaSes, as it simply controls how long to wait
before the sshd server sends a message to a quiet client asking for a
response.

For more information, see:
https://learn.microsoft.com/en-us/azure/security/fundamentals/azure-marketplace-images#linux-and-open-source-os-images