Stop using ubuntu release name in filenames
Now that different stemcell releases are on separate branches we do not
need to differentiate filenames by os release name.

This commit also removes a number of non Ubuntu Jammy files, there are
likely more that could be cleaned up, notably some comments /
conditionals in the following files:

- stemcell_builder/stages/system_kernel/apply.sh
- stemcell_builder/stages/rsyslog_config/assets/rsyslog_logrotate.conf

The contents of `bosh-stemcell/image-metalinks/#ubuntu-jammy/` should
probably also be examined to see if they can be deleted and/or consumed
from some other branch of this repo.