Move the stemcell acceptance specs to top-level
Prior to the go module dependency management tooling Golang source code needed
to be in a repository-relative path under `src/`, this is no longer true. This
commit moves `src/github.com/cloudfoundry/stemcell-acceptance-tests/` to the
top-level of the repo, renaming the folder to `acceptance-tests/`. The golang
packages are updated to reflect this.

In addition the `tools` package is removed since Ginkgo v2 no longer requires
this to ensure the CLI code is available when testing.