Decreases "recordfail" timeout to five seconds
This commit reduces the boot timeout for the "recordfail" GRUB boot
logic from its default (30 seconds) to five seconds. The ordinary boot
wait time is five seconds, so this makes the boot wait the same
regardless of which branch is taken.

Currently, we strongly believe that the first boot of any VM based on a
stemcell falls into the "recordfail" branch, and has its startup delayed
by 30 seconds for no good reason.

Do note that this environment variable is not documented in the official
GRUB documentation, but -then again- most of the environment variables
used by the GRUB configuration generation machinery are also
undocumented... source code inspection is the only way to learn of their
existence. The relevant files for this commit are the grub-mkconfig
script (which has a list of user-provided environment variables that
influence its template rendering), and the /etc/grub.d/00_header
template, which is currently responsible for emitting the timeout
sections of the GRUB config file.

Signed-off-by: Rajan Agaskar <rajan.agaskar@broadcom.com>