fix: run logrotate once every 5 minutes
When a system is processing lots of data producing a high volume of logs
running logrotate only once every 15 mintues can cause the disk to fill
up if it is not big enough. This commit changes the logrotate timer to
fire once per minute. If no logs need to be rotated logrotate will exit
without modifying anything, this can be controlled via custom logrotate
configuration.

This commit also removes the OnBootSec property as the timer fires every
5 minutes anyway, RandomizedDelaySec is reduced to one minute.

Resolves: https://github.com/cloudfoundry/bosh-linux-stemcell-builder/issues/424