This PR allows stepping system clock on any clock update (#422)
V-260520
SRG-OS-000356-GPOS-00144
SV-260520r1044776_rule
CAT III
UBTU-22-252015
Ubuntu 22.04 LTS must synchronize internal information system clocks to the authoritative time source when the time difference is greater than one second.

 Inaccurate time stamps make it more difficult to correlate events and can lead to an inaccurate analysis. Determining the correct time a particular event occurred on a system is critical when conducting forensic analysis and investigating system events.

Synchronizing internal information system clocks provides uniformity of time stamps for information systems with multiple system clocks and systems connected over a network. Organizations should consider setting time periods for different types of systems (e.g., financial, legal, or mission-critical systems).

Organizations should also consider endpoints that may not have regular access to the authoritative time server (e.g., mobile, teleworking, and tactical endpoints). This requirement is related to the comparison done every 24 hours in SRG-OS-000355 because a comparison must be done to determine the time difference.

Verify Ubuntu 22.04 LTS synchronizes internal system clocks to the authoritative time source when the time difference is greater than one second.

**Note**: If the system is not networked, this requirement is not applicable.

Check the value of "makestep" by using the following command:

     $ grep -ir makestep /etc/chrony*
     makestep 1 -1

If `makestep` is not set to `1 -1`, is commented out, or is missing, this is a finding.

Verify the NTP service is active and the system clock is synchronized with the authoritative time source:

     $ timedatectl | grep -Ei '(synchronized|service)'
     System clock synchronized: yes
     NTP service: active

If the NTP service is not active, this is a finding.

If the system clock is not synchronized, this is a finding.

Configure chrony to synchronize the internal system clocks to the authoritative source when the time difference is greater than one second by doing the following:

Edit the `/etc/chrony/chrony.conf` file and add:

     makestep 1 -1

Restart the chrony service:

     $ sudo systemctl restart chrony.service

*Normally, it is recommended to allow the step only in the first few updates, but in some cases (e.g. a computer without an RTC or virtual machine which can be suspended and resumed with an incorrect time) it might be necessary to allow the step on any clock update.*

[3.4. Is chronyd allowed to step the system clock?](https://chrony-project.org/faq.html#_is_chronyd_allowed_to_step_the_system_clock)

[Kubernetes image-builder](https://github.com/kubernetes-sigs/image-builder/blob/main/images/capi/ansible/roles/providers/tasks/azure.yml#L26-L32)